refactor(cli)!: relocate generate business logic to _app/ (transport-neutral)
Relocate the transport-neutral `generate` business logic out of the three
`cli/services/generate*` modules into `src/notebooklm/_app/`, mirroring the
merged download / source-mutations precedent. The CLI service modules shrink to
thin adapters/re-exports; `cli/generate_cmd.py` (and its 59 `patch(
"...generate_cmd.NotebookLMClient")` seams) is untouched, so all Click
adaptation stays in the command layer.

New neutral modules (split three ways to stay under the ADR-0008 module-size
budget, mirroring the original CLI service trio):

* `_app/generate_plans.py` — enum/format maps, `GenerationPlan` /
  `GenerationKind` / `GenerationPlanValidationError`, `build_generation_plan`
  + per-kind builders. `parameter_explicit` / `language_resolver` stay injected
  (the Click `ParameterSource.COMMANDLINE` probe + the env/config language chain
  are the adapter's). `GenerationPlanValidationError` now subclasses
  `notebooklm.exceptions.ValidationError` so `_app.errors.classify` covers it
  as VALIDATION uniformly (the unknown-kind raise routes through it too — no
  bare `ValueError` escapes the `_app` boundary).
* `_app/generate_retry.py` — retry-with-backoff loop, wait-for-completion
  orchestration, `GenerationOutcome`, status extractors, spinner status-line
  formatter. `wait_context` / `wait_start_sink` are neutral injected seams.
* `_app/generate.py` — the `execute_generation` dispatcher + the typed
  `GenerationExecutionResult`; re-exports the plan/retry surface so
  `_app.generate` is the single import point. The notebook/source resolvers are
  injected (the CLI passes `cli.resolve.resolve_notebook_id` /
  `resolve_source_ids`, whose full-id fast paths preserve the recorded RPC call
  set, so the cassettes still match without re-recording).

The CLI `execute_generation` reads the `cli.resolve` resolvers at call time, so
the historical `monkeypatch.setattr(resolve_module, "resolve_notebook_id", ...)`
seam keeps landing. The `--json` envelope + exit-code rendering stays in
`cli/generate_cmd.py` (the typed result -> envelope renderer is unchanged), so
`--json` output is byte-stable. `_app` results carry no `.payload` /
envelope-builders — typed fields only.
