+++ src/notebooklm/_app/__init__.py
@@ -48,6 +48,18 @@ from .download import (
 )
 from .errors import ClassifiedError, ErrorCategory, classify
 from .events import ProgressEvent, ProgressSink
+from .generate import (
+    GenerationExecutionResult,
+    GenerationKind,
+    GenerationOutcome,
+    GenerationPlan,
+    GenerationPlanValidationError,
+    build_generation_plan,
+    execute_generation,
+    generate_with_retry,
+    generation_outcome_from_status,
+    handle_generation_result,
+)
 from .resolve import AmbiguousIdError, Resolution, resolve_ref, validate_id
 from .serialize import to_jsonable
 from .source_add import (
@@ -137,6 +149,17 @@ __all__ = [
     "validate_id",
     "ProgressEvent",
     "ProgressSink",
+    # generate
+    "GenerationExecutionResult",
+    "GenerationKind",
+    "GenerationOutcome",
+    "GenerationPlan",
+    "GenerationPlanValidationError",
+    "build_generation_plan",
+    "execute_generation",
+    "generate_with_retry",
+    "generation_outcome_from_status",
+    "handle_generation_result",
     # download
     "FORMAT_EXTENSIONS",
     "ArtifactDict",
